Flask-App/requirements.txt
Flask==2.2.2
tensorflow==2.14.0
pillow==10.0.1